+++ CHANGELOG.md
@@ -11,9 +11,20 @@ All notable changes to this project will be documented in this file, in reverse 
   the `DateTime` constructor will still result in a `DateTime` instance during
   hydration, even if the string does not follow the provided date-time format.
 
+- [#14](https://github.com/zendframework/zend-hydrator/pull/14) adds the following `final` classes:
+  - `\Zend\Hydrator\NamingStrategy\UnderscoreNamingStrategy\UnderscoreToCamelCaseFilter`
+  - `\Zend\Hydrator\NamingStrategy\UnderscoreNamingStrategy\CamelCaseToUnderscoreFilter`
+
 ### Changed
 
-- Nothing.
+- [#14](https://github.com/zendframework/zend-hydrator/pull/14) replaces usage of zend-filter with the hardcoded filters referenced in
+  the above section.
+
+- [#14](https://github.com/zendframework/zend-hydrator/pull/14) made the following visibility changes to `\Zend\Hydrator\NamingStrategy\UnderscoreNamingStrategy`:
+  - static property `$underscoreToStudlyCaseFilter` was renamed to `$underscoreToCamelCaseFilter` and marked `private`
+  - static property `$camelCaseToUnderscoreFilter` was marked `private`
+  - method `getCamelCaseToUnderscoreFilter` was marked `private`
+  - method `getUnderscoreToStudlyCaseFilter` was renamed to `getUnderscoreToCamelCaseFilter` and marked `private`
 
 ### Deprecated